requirements.txt
requests==2.27.1
GitPython==3.1.27